requirements.txt
ctransformers
sentence-transformers
pinecone
langchain-community
langchain-pinecone
langchain
flask
pypdf
python-dotenv
-e .